Improved DB initialization + runtime performance. Patched routing bug.
All DB's are synced as codeoil server launches and no more syncing occurs as the
application runs. Rewrote some of the column names to be more representative of
what they contain.

Last but not least I put in some error-checking logic to confirm that only integers
are passed as challenge ID's. I still need to verify that the integer passed is a
valid challenge but this is a step in the right direction.